Fixed Inventory Component, modularizing it. Created Constants file to fix magic numbers. Fixed extra rankcomponent
