+++ MiniChallenge05/MiniChallenge05 Watch App/Utility/Constants.swift
@@ -1,0 +1,16 @@
+//
+//  Constants.swift
+//  MiniChallenge05 Watch App
+//
+//  Created by Luca on 04/06/24.
+//
+
+import Foundation
+import SwiftUI
+
+struct Constants {
+    
+    static let itemWidth = WKInterfaceDevice.current().screenBounds.width/6.5
+    static let itemHeight = WKInterfaceDevice.current().screenBounds.width/6.5
+    
+}
