+++ MiniChallenge05/MiniChallenge05 Watch App/Views/AgentProgressView.swift
@@ -18,8 +18,6 @@ struct AgentProgressView: View {
                         RankComponent()
                     }.buttonStyle(CustomButton())
 
-                    RankComponent()
-                        .padding(.vertical)
                     MissionCounterComponent()
                         .padding(.vertical)
                     TotalDistanceComponent()
